feat: jwt authentication

+++ src/resolvers/product.resolver.js
@@ -3,6 +3,7 @@ import { deleteProduct } from "../db/database_functions/product/deleteProduct.js
 import { updateProduct } from "../db/database_functions/product/updateProduct.js";
 import { getProducts } from "../db/database_functions/product/getProducts.js";
 import { getProduct } from "../db/database_functions/product/getProduct.js";
+import { CustomError, customGraphqlErrorCodes, httpStatusCodes } from "../constants/constants.js";
 
 export const productResolvers = {
   Query: {
@@ -24,24 +25,39 @@ export const productResolvers = {
     },
   },
   Mutation: {
-    async createProduct(_, { name, price, user }) {
+    async createProduct(_, { name, price, user }, { userId }) {
       try {
+        if (!userId) throw new CustomError(httpStatusCodes['Unauthorized'], customGraphqlErrorCodes['UNAUTHORIZED'], 'Unauthorized');
+
+        if (userId !== user) throw new CustomError(httpStatusCodes['Forbidden'], customGraphqlErrorCodes['FORBIDDEN'], 'Forbidden');
+
         const product = await createProduct({ name, price, user })
         return product;
       } catch (error) {
         throw error;
       }
     },
-    async updateProduct(_, { _id, ...dataToUpdate }) {
+    async updateProduct(_, { _id, ...dataToUpdate }, { userId }) {
       try {
-        const product = await updateProduct(_id, dataToUpdate);
-        return product;
+        if (!userId) throw new CustomError(httpStatusCodes['Unauthorized'], customGraphqlErrorCodes['UNAUTHORIZED'], 'Unauthorized');
+        const product = await getProduct(_id);
+
+        if (product?.user?.toString() !== userId) throw new CustomError(httpStatusCodes['Forbidden'], customGraphqlErrorCodes['FORBIDDEN'], 'Forbidden');
+
+        const updatedProduct = await updateProduct(_id, dataToUpdate);
+        return updatedProduct;
       } catch (error) {
         throw error;
       }
     },
-    async deleteProduct(_, { _id }) {
+    async deleteProduct(_, { _id }, { userId }) {
       try {
+        if (!userId) throw new CustomError(httpStatusCodes['Unauthorized'], customGraphqlErrorCodes['UNAUTHORIZED'], 'Unauthorized');
+
+        const product = await getProduct(_id);
+
+        if (product?.user?.toString() !== userId) throw new CustomError(httpStatusCodes['Forbidden'], customGraphqlErrorCodes['FORBIDDEN'], 'Forbidden');
+
         await deleteProduct(_id);
         return "Product deleted successfully!"
       } catch (error) {
